Rename IErrorMessageFormatter descendants to correct FrankDoc
